coven of the meek implementation

+++ jokers/joker_utils.lua
@@ -1,8 +1,14 @@
 calc_spell_cast = function(self, card, context)
     if context.after and not context.blueprint then
-        SMODS.destroy_cards(card, nil, nil, true)
-        return {
-            message = localize('wlt_cast_ex')
-        }
+        if card.ability.extra.fresh then
+            -- Solves some race conditions so spells don't immediately
+            -- destroy themselves if they are created in context.after
+            card.ability.extra.fresh = false
+        else
+            SMODS.destroy_cards(card, nil, nil, true)
+            return {
+                message = localize('wlt_cast_ex')
+            }
+        end
     end
 end
